Teleportation Select & Preview
Added a system to allow for teleporting to set positon by placing prefabs in those locations.
When the player starts aiming, all the teleport points will reveal themselves. When the player stops aiming these points hide again. (uses Bakr's event system)
When a player is aiming and aims at a teleport point, a simple particle system plays to show it is selected. The particles stop when you stop aiming at it