Move Validty Check & Puzzle Solved
Added a validty check for the movement of the grid cursor. This happens before the cursor's position is changed and check the following conditions (in order):
1. Is the puzzle already solved
2. Do we move out of the vertical bounds (negatively then positevly)
3. Do we move out of the horizontal bounds (negatively then positevly)
4. Does the tile we would move to exist
4.1 if it does, is it already full

Changed the way we check for if the puzzle is solved a bit. We loop through all the tiles in the grid, add the pathability of all the tiles with type "end" and return whether all those tiles are full (ie. the inverse of if there are any false values in the list).
Added a UnityEvent that gets invoked when the puzzle is solved.